Pull Request from decimals to main

+++ calc.py
@@ -1,20 +1,20 @@
 # Perform simple arithmetic encoded in an input string:
 # '1 + 2' -> 3,
 # '1 - 2' -> -1,
-# '1 * 2' -> 2
+# '1 * 2' -> 2,
 # '1 / 2' -> 0.5
 
 def compute(expression):
-    num0, operator, num1 = expression.split(' ')
-    num0, num1 = int(num0), int(num1)
-    if operator == '+':
-        return num0 + num1
-    elif operator == '-':
-        return num0 - num1
-    elif operator == '*':
-        return num0 * num1
-    elif operator == '/':
-        return num0 / num1
-    else:
-        print('unknown operator!')
-        return None
+  num0, operator, num1 = expression.split(' ')
+  num0, num1 = float(num0), float(num1)
+  if operator == '+':
+    return num0 + num1
+  elif operator == '-':
+    return num0 - num1
+  elif operator == '*':
+    return num0 * num1
+  elif operator == '/':
+    return num0 / num1
+  else:
+    print('unknown operator!')
+    return None
